ml-service/requirements.txt
fastapi==0.110.0
uvicorn==0.27.1
torch==2.2.1
torchvision==0.17.1
pillow==10.2.0
python-multipart==0.0.9
numpy==1.26.4
scikit-learn==1.4.1.post1
mlflow==2.11.1
prometheus-client==0.20.0
